Fixed typo's in sodium.i that caused crashes in the Java vm for *_state related datatypes.
Added null-pointer checks on byte arrays in the JNI code to avoid nullpointerexceptions in code where null pointers are valid (such as a null key in blake2b).

Some manual tests using the new stream-api's, no crashes. Correctness of output to be verified.

TODO: should make sure that JNI calls to GetByteArrayElements always point to the raw memory, not a copy.